+++ src/birdnet/models/v2m4/model_v2m4_tflite.py
@@ -8,7 +8,7 @@ import numpy.typing as npt
 from ordered_set import OrderedSet
 from tensorflow.lite.python.interpreter import Interpreter
 
-from birdnet.models.model_v2m4_base import (AVAILABLE_LANGUAGES, AudioModelBaseV2M4,
+from birdnet.models.v2m4.model_v2m4_base import (AVAILABLE_LANGUAGES, AudioModelBaseV2M4,
                                             MetaModelBaseV2M4, get_internal_version_app_data_folder,
                                             validate_language)
 from birdnet.types import Language