+++ src/birdnet_tests/models/m2v4/model_v2m4_protobuf_py/test_predict_species_at_location_and_time_gpu.py
@@ -6,10 +6,10 @@ import pytest
 import tensorflow as tf
 from tqdm import tqdm
 
-from birdnet.models.model_v2m4_protobuf import MetaModelV2M4Protobuf
+from birdnet.models.v2m4.model_v2m4_protobuf import MetaModelV2M4Protobuf
 from birdnet.types import SpeciesPrediction
 from birdnet_tests.helper import TEST_RESULTS_DIR, species_prediction_is_equal
-from birdnet_tests.models.test_predict_species_at_location_and_time import (
+from birdnet_tests.models.m2v4.test_predict_species_at_location_and_time import (
   LocationTestCase, create_ground_truth_test_file,
   model_test_identical_predictions_return_same_result, model_test_no_week,
   model_test_predictions_are_globally_correct, model_test_using_no_threshold_returns_all_species,