+++ src/birdnet_tests/models/m2v4/model_v2m4_raven_custom_py/test_custom_audio_model_v2m4_raven.py
@@ -2,8 +2,8 @@
 import numpy.testing as npt
 import pytest
 
-from birdnet.models.model_v2m4_raven_custom import CustomAudioModelV2M4Raven
-from birdnet.models.model_v2m4_tflite_custom import CustomAudioModelV2M4TFLite
+from birdnet.models.v2m4.model_v2m4_raven_custom import CustomAudioModelV2M4Raven
+from birdnet.models.v2m4.model_v2m4_tflite_custom import CustomAudioModelV2M4TFLite
 from birdnet_tests.helper import TEST_FILE_WAV, TEST_FILES_DIR, convert_predictions_to_numpy
 
 CLASSIFIER_FOLDER = TEST_FILES_DIR / "v2m4" / "custom_model_raven"